FIX: Order definition for Header class
